[MNT] removed support for local processing, added tests, refactored methods

+++ func_preprocess/workflows.py
@@ -3,6 +3,7 @@
 run_preproc : coordinate methods for preprocessing workflow
 
 """
+
 import os
 import shutil
 from func_preprocess import preprocess, helper_tools
@@ -14,17 +15,11 @@ def run_preproc(
     proj_raw,
     proj_deriv,
     work_deriv,
-    sing_fmriprep,
-    tplflow_dir,
-    fs_license,
     fd_thresh,
     ignore_fmaps,
-    sing_afni,
     log_dir,
-    run_local,
-    user_name=None,
-    rsa_key=None,
     keoki_path="/mnt/keoki/experiments2/EmoRep/Exp2_Compute_Emotion/data_scanner_BIDS",  # noqa: E501
+    test_mode=False,
 ):
     """Functional preprocessing pipeline for EmoRep.
 
@@ -34,37 +29,22 @@ def run_preproc(
         BIDS subject identifier
     sess_list : list
         BIDS session identifiers
-    proj_raw : path
-        Location of project rawdata, e.g.
-        /hpc/group/labarlab/EmoRep_BIDS/rawdata
-    proj_deriv : path
-        Location of project derivatives, e.g.
-        /hpc/group/labarlab/EmoRep_BIDS/derivatives
-    work_deriv : path
-        Location of work derivatives, e.g.
-        /work/foo/EmoRep_BIDS/derivatives
-    sing_fmriprep : path, str
-        Location of fmiprep singularity image
-    tplflow_dir : path, str
-        Clone location of templateflow
-    fs_license : path, str
-        Location of FreeSurfer license
+    proj_raw : str, os.PathLike
+        Location of project rawdata
+    proj_deriv : str, os.PathLike
+        Location of project derivatives
+    work_deriv : str, os.PathLike
+        Location of work derivatives
     fd_thresh : float
         Threshold for framewise displacement
     ignore_fmaps : bool
         Whether to incorporate fmaps in preprocessing
-    sing_afni : path, str
-        Location of afni singularity iamge
-    log_dir : path
+    log_dir : str, os.PathLike
         Location for writing logs
-    run_local : bool
-        Whether job, subprocesses are run locally
-    user_name : str, optional
-        User name for DCC, labarserv2
-    rsa_key : str, os.PathLike, optional
-        Location of RSA key for labarserv2
     keoki_path : str, os.PathLike, optional
         Location of project directory on Keoki
+    test_mode : bool, optional
+        Used to avoid data push, cleanup during testing
 
     Raises
     ------
@@ -72,46 +52,34 @@ def run_preproc(
         Unexpected types for passed args
 
     """
-    # Check passed args
+    # Check args
     if not isinstance(sess_list, list):
-        raise TypeError("Expected type list for sess_list")
-    for _chk_bool in [ignore_fmaps, run_local]:
-        if not isinstance(_chk_bool, bool):
-            raise TypeError(
-                "Expected bool type for options : --ignore_fmaps, "
-                + "--run_local"
-            )
+        raise TypeError("Expected type 'list' for sess_list")
+    if not isinstance(ignore_fmaps, bool):
+        raise TypeError("Expected type 'bool' for ignore_fmaps")
     if not isinstance(fd_thresh, float):
-        raise TypeError("Expected float type for --fd_thresh")
-    if (not run_local and user_name is None) or (
-        not run_local and rsa_key is None
-    ):
-        raise ValueError("user name and rsa key required on DCC")
+        raise TypeError("Expected type 'float' for fd_thresh")
+
+    # Check env
+    helper_tools.check_env()
 
     # Download needed files
-    if not run_local:
-        sync_data = helper_tools.PullPush(
-            os.path.dirname(proj_raw), log_dir, user_name, rsa_key, keoki_path
-        )
-        for sess in sess_list:
-            sync_data.pull_rawdata(subj, sess)
+    sync_data = helper_tools.PullPush(
+        os.path.dirname(proj_raw), log_dir, keoki_path
+    )
+    for sess in sess_list:
+        _ = sync_data.pull_rawdata(subj, sess)
 
     # Run FreeSurfer, fMRIPrep
-    run_fs = preprocess.RunFreeSurfer(
-        subj, proj_raw, work_deriv, log_dir, run_local
-    )
+    run_fs = preprocess.RunFreeSurfer(subj, proj_raw, work_deriv, log_dir)
     run_fs.recon_all(sess_list)
     run_fp = preprocess.RunFmriprep(
         subj,
         proj_raw,
         work_deriv,
-        sing_fmriprep,
-        tplflow_dir,
-        fs_license,
         fd_thresh,
         ignore_fmaps,
         log_dir,
-        run_local,
     )
     run_fp.fmriprep(sess_list)
     fp_dict = run_fp.get_output()
@@ -120,16 +88,11 @@ def run_preproc(
     _ = preprocess.fsl_preproc(
         work_deriv,
         fp_dict,
-        sing_afni,
         subj,
         log_dir,
-        run_local,
     )
 
-    # Clean up only dcc
-    if run_local:
-        return
-
+    # Copy files from work to group partition
     helper_tools.copy_clean(
         subj,
         sess_list,
@@ -137,5 +100,9 @@ def run_preproc(
         work_deriv,
         log_dir,
     )
+
+    # Send data to keoki and clean up
+    if test_mode:
+        return
     sync_data.push_derivatives(sess_list)
     shutil.rmtree(os.path.join(proj_raw, subj))
